Split utils module into public utils and private _io_helpers; fix remaining typing issues; add some to-do's.
